test: reorganize jest tests
