+++ package.json
@@ -3,18 +3,6 @@
   "private": true,
   "version": "1.0.0",
   "description": "frontend for the sextant cluster creation service",
-  "jest": {
-    "verbose": true,
-    "reporters": [
-      "default",
-      [
-        "jest-tap-reporter",
-        {
-          "filePath": "build/results.tap"
-        }
-      ]
-    ]
-  },
   "scripts": {
     "develop": "webpack serve --config config/webpack.dev.js",
     "build-test": "webpack --config config/webpackTest.js",
@@ -83,6 +71,7 @@
     "@babel/preset-react": "^7.18.6",
     "@types/react-dom": "^18.0.11",
     "@types/webpack-env": "^1.18.0",
+    "babel-jest": "^29.5.0",
     "babel-loader": "^9.1.2",
     "babel-tape-runner": "^3.0.0",
     "clean-webpack-plugin": "^3.0.0",
@@ -96,9 +85,10 @@
     "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "html-webpack-plugin": "^5.5.0",
-    "jest": "^27.1.0",
-    "jest-tap-reporter": "^1.3.0",
+    "jest": "^29.5.0",
+    "jest-tap-reporter": "^1.9.0",
     "style-loader": "^3.3.2",
+    "ts-jest": "^29.0.5",
     "ts-loader": "^9.4.2",
     "typescript": "^5.0.2",
     "webpack": "^5.76.2",
